+++ extensions/projects/src/cache/all-projects-cache-builder.ts
@@ -3,15 +3,43 @@ import { exec } from "child_process";
 import { promisify } from "util";
 import { ApplicationCache } from "./application-cache";
 import { CacheType, ProjectType, SourceRepo } from "../types";
+import { v4 as uuidv4 } from "uuid";
 const execp = promisify(exec);
 
-export async function buildAllProjectsCache(paths: string[], maxDepth: number): Promise<SourceRepo[]> {
+export async function buildAllProjectsCache(paths: string[], projectTypes: ProjectType[]): Promise<SourceRepo[]> {
   const allProjectsCache = new ApplicationCache(CacheType.ALL_PROJECTS);
   let foundRepos: SourceRepo[] = [];
   await Promise.allSettled(
     paths.map(async (path) => {
-      const findCmd = `find -L ${path} -maxdepth ${maxDepth} \\( -name 'package.json' -o -name 'pom.xml' -o -name 'build.gradle' \\) -type f -not -path "*/node_modules/*"`;
-      const { stdout, stderr } = await execp(findCmd);
+      let spotlightSearchParameters: string[] = [];
+
+      if (projectTypes.length == 0 || projectTypes.includes(ProjectType.XCODE)) {
+        spotlightSearchParameters = [
+          ...spotlightSearchParameters,
+          "kMDItemDisplayName == *.xcodeproj",
+          "kMDItemDisplayName == *.xcworkspace",
+          "kMDItemDisplayName == Package.swift",
+          "kMDItemDisplayName == *.playground",
+        ];
+      }
+
+      if (projectTypes.length == 0 || projectTypes.includes(ProjectType.GRADLE)) {
+        spotlightSearchParameters = [...spotlightSearchParameters, "kMDItemDisplayName == build.gradle"];
+      }
+
+      if (projectTypes.length == 0 || projectTypes.includes(ProjectType.MAVEN)) {
+        spotlightSearchParameters = [...spotlightSearchParameters, "kMDItemDisplayName == pom.xml"];
+      }
+
+      if (projectTypes.length == 0 || projectTypes.includes(ProjectType.NODE)) {
+        spotlightSearchParameters = [...spotlightSearchParameters, "kMDItemDisplayName == package.json"];
+      }
+
+      // Execute command
+      const { stdout, stderr } = await execp(
+        `mdfind -onlyin ${path} '${spotlightSearchParameters.join(" || ")}' | grep -v "node_modules\\|META-INF"`
+      );
+
       if (stderr) {
         showToast(Toast.Style.Failure, "Find Failed", stderr);
         console.error(`error: ${stderr}`);
@@ -41,10 +69,12 @@ export async function buildAllProjectsCache(paths: string[], maxDepth: number): 
 
 function parseRepoPaths(repoPaths: string[]): SourceRepo[] {
   return repoPaths.map((path) => {
+    const uuid = uuidv4();
     if (path.endsWith("package.json")) {
       const fullPath = path.replace("/package.json", "");
       const name = fullPath.split("/").pop() ?? "unknown";
       return {
+        id: uuid,
         name: name,
         icon: "node-js.png",
         fullPath: fullPath,
@@ -54,6 +84,7 @@ function parseRepoPaths(repoPaths: string[]): SourceRepo[] {
       const fullPath = path.replace("/pom.xml", "");
       const name = fullPath.split("/").pop() ?? "unknown";
       return {
+        id: uuid,
         name: name,
         icon: "maven.png",
         fullPath: fullPath,
@@ -63,14 +94,35 @@ function parseRepoPaths(repoPaths: string[]): SourceRepo[] {
       const fullPath = path.replace("/build.gradle", "");
       const name = fullPath.split("/").pop() ?? "unknown";
       return {
+        id: uuid,
         name: name,
         icon: "gradle.png",
         fullPath: fullPath,
         type: ProjectType.GRADLE,
       };
+    } else if (path.endsWith(".xcodeproj") || path.endsWith(".xcworkspace") || path.endsWith("Package.swift")) {
+      const fullPath = path.substring(0, path.lastIndexOf("/"));
+      const name = path.split("/").pop() ?? "unknown";
+      return {
+        id: uuid,
+        name: name,
+        icon: "xcode.png",
+        fullPath: fullPath,
+        type: ProjectType.XCODE,
+      };
+    } else if (path.endsWith(".playground")) {
+      const name = path.split("/").pop() ?? "unknown";
+      return {
+        id: uuid,
+        name: name,
+        icon: "xcode.png",
+        fullPath: path,
+        type: ProjectType.XCODE,
+      };
     }
 
     return {
+      id: uuid,
       name: "unknown",
       icon: "unknown.png",
       fullPath: path,