+++ extensions/projects/src/common-utils.ts
@@ -1,5 +1,5 @@
 import { SourceRepo } from "./types";
 
 export function getRepoKey(repo: SourceRepo): string {
-  return `${repo.fullPath}~${repo.type}`;
+  return `${repo.fullPath}/${repo.name}~${repo.type}`;
 }