+++ extensions/projects/src/types.ts
@@ -1,4 +1,5 @@
 export interface SourceRepo {
+  id: string;
   name: string;
   fullPath: string;
   icon: string;
@@ -19,8 +20,9 @@ export interface OpenWith {
 
 export interface Preferences {
   repoScanPath: string;
-  repoScanDepth?: number;
+  scanForProjectTypes: string;
   openNodeWith: OpenWith;
+  openXcodeWith: OpenWith;
   openMavenWith: OpenWith;
   openGradleWith: OpenWith;
   openWith1: OpenWith;
@@ -32,6 +34,7 @@ export enum ProjectType {
   NODE = "node",
   MAVEN = "maven",
   GRADLE = "gradle",
+  XCODE = "xcode",
   UNKNOWN = "unknown",
 }