+++ extensions/projects/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Projects Changelog
 
+## [Added support for Xcode projects + Advanced Search] - 2022-05-16
+
+- Faster search with mfind (instead of find)
+- Supports Xcode projects
+- Open in finder & terminal
+- Better defaults for project types
+- Advanced search support - (type, dir)
+
 ## [Added node, maven & gradle projects support] - 2022-05-12
 
 - Initial version